+++ Math.cs
@@ -20,31 +20,31 @@ namespace Electron2D
         /// This value is different from System.Double.Epsilon, which is the smallest, positive, denormalized f64/double.
 		/// </para>
         /// </summary>
-		public const double Epsilon = Double.Epsilon;
+		public const float Epsilon = (float)Double.Epsilon;
 
 		/// <summary>
         /// Double precision constant for positive infinity.
         /// </summary>
-        public const double Infinity = Double.PositiveInfinity;
+        public const float Infinity = (float)Double.PositiveInfinity;
 
 		/// <summary>
         /// Double precision constant for negative infinity.
         /// </summary>
-        public const double NegativeInfinity = Double.NegativeInfinity;
+        public const float NegativeInfinity = (float)Double.NegativeInfinity;
 
 		/// <summary>The mathematical constant pi. Approximately 3.14. This is a f64/double precision constant.</summary>
-        public const double PI = System.Math.PI;
+        public const float PI = (float)System.Math.PI;
 
 		/// <summary>Radians-to-degrees conversion constant.</summary>
-		public static double Rad2Deg = 360 / (PI * 2);
+		public static float Rad2Deg = 360F / (PI * 2F);
 
 		/// <summary>Degrees-to-radians conversion constant.</summary>
-		public static double Deg2Rad = (PI * 2) / 360;
+		public static float Deg2Rad = (PI * 2F) / 360F;
 
-		public static double Sin(double f) { return System.Math.Sin(f); }
+		public static float Sin(float f) { return (float)System.Math.Sin(f); }
 
         // Returns the cosine of angle /f/ in radians.
-        public static double Cos(double f) { return System.Math.Cos(f); }
+        public static float Cos(float f) { return (float)System.Math.Cos(f); }
 
         // Returns the tangent of angle /f/ in radians.
         public static double Tan(double f) { return System.Math.Tan(f); }
@@ -62,10 +62,10 @@ namespace Electron2D
         public static double Atan2(double y, double x) { return System.Math.Atan2(y, x); }
 
         // Returns square root of /f/.
-        public static double Sqrt(double f) { return System.Math.Sqrt(f); }
+        public static float Sqrt(float f) { return (float)System.Math.Sqrt(f); }
 
         // Returns the absolute value of /f/.
-        public static double Abs(double f) { return System.Math.Abs(f); }
+        public static float Abs(float f) { return (float)System.Math.Abs(f); }
 
         // Returns the absolute value of /value/.
         public static int Abs(int value) { return System.Math.Abs(value); }
@@ -101,14 +101,14 @@ namespace Electron2D
             }
             return m;
         }
-        public static double Max(double a, double b) { return a > b ? a : b; }
+        public static float Max(float a, float b) { return a > b ? a : b; }
         // Returns largest of two or more values.
-        public static double Max(params double[] values)
+        public static float Max(params float[] values)
         {
             int len = values.Length;
             if (len == 0)
                 return 0;
-            double m = values[0];
+            float m = values[0];
             for (int i = 1; i < len; i++)
             {
                 if (values[i] > m)
@@ -134,43 +134,43 @@ namespace Electron2D
         }
 
         // Returns /f/ raised to power /p/.
-        public static double Pow(double f, double p) { return System.Math.Pow(f, p); }
+        public static float Pow(float f, float p) { return (float)System.Math.Pow(f, p); }
 
         // Returns e raised to the specified power.
-        public static double Exp(double power) { return System.Math.Exp(power); }
+        public static float Exp(float power) { return (float)System.Math.Exp(power); }
 
         // Returns the logarithm of a specified number in a specified base.
-        public static double Log(double f, double p) { return System.Math.Log(f, p); }
+        public static float Log(float f, float p) { return (float)System.Math.Log(f, p); }
 
         // Returns the natural (base e) logarithm of a specified number.
-        public static double Log(double f) { return System.Math.Log(f); }
+        public static float Log(float f) { return (float)System.Math.Log(f); }
 
         // Returns the base 10 logarithm of a specified number.
-        public static double Log10(double f) { return System.Math.Log10(f); }
+        public static float Log10(float f) { return (float)System.Math.Log10(f); }
 
         // Returns the smallest integer greater to or equal to /f/.
-        public static double Ceil(double f) { return System.Math.Ceiling(f); }
+        public static float Ceil(float f) { return (float)System.Math.Ceiling(f); }
 
         // Returns the largest integer smaller to or equal to /f/.
-        public static double Floor(double f) { return System.Math.Floor(f); }
+        public static float Floor(float f) { return (float)System.Math.Floor(f); }
 
         // Returns /f/ rounded to the nearest integer.
-        public static double Round(double f) { return System.Math.Round(f); }
+        public static float Round(float f) { return (float)System.Math.Round(f); }
 
         // Returns the smallest integer greater to or equal to /f/.
-        public static int CeilToInt(double f) { return (int)System.Math.Ceiling(f); }
+        public static int CeilToInt(float f) { return (int)System.Math.Ceiling(f); }
 
         // Returns the largest integer smaller to or equal to /f/.
-        public static int FloorToInt(double f) { return (int)Math.Floor(f); }
+        public static int FloorToInt(float f) { return (int)Math.Floor(f); }
 
         // Returns /f/ rounded to the nearest integer.
-        public static int RoundToInt(double f) { return (int)Math.Round(f); }
+        public static int RoundToInt(float f) { return (int)Math.Round(f); }
 
         // Returns the sign of /f/.
-        public static double Sign(double f) { return f >= 0 ? 1 : -1; }
+        public static float Sign(float f) { return f >= 0 ? 1 : -1; }
 
         // Clamps a value between a minimum float and maximum float value.
-        public static double Clamp(double value, double min, double max)
+        public static float Clamp(float value, float min, float max)
         {
             if (value < min)
                 return min;
@@ -193,7 +193,7 @@ namespace Electron2D
         }
 
         // Clamps value between 0 and 1 and returns value
-        public static double Clamp01(double value)
+        public static float Clamp01(float value)
         {
             if (value < 0)
                 return 0;
@@ -204,28 +204,28 @@ namespace Electron2D
         }
 
         // Interpolates between /a/ and /b/ by /t/. /t/ is clamped between 0 and 1.
-        public static double Lerp(double a, double b, double t)
+        public static float Lerp(float a, float b, float t)
         {
             return a + ((b - a) * Clamp01(t));
         }
 
         // Interpolates between /a/ and /b/ by /t/ without clamping the interpolant.
-        public static double LerpUnclamped(double a, double b, double t)
+        public static float LerpUnclamped(float a, float b, float t)
         {
             return a + ((b - a) * t);
         }
 
         // Same as ::ref::Lerp but makes sure the values interpolate correctly when they wrap around 360 degrees.
-        public static double LerpAngle(double a, double b, double t)
+        public static float LerpAngle(float a, float b, float t)
         {
-            double delta = Repeat((b - a), 360);
+            float delta = Repeat((b - a), 360);
             if (delta > 180)
                 delta -= 360;
             return a + (delta * Clamp01(t));
         }
 
         // Moves a value /current/ towards /target/.
-        static public double MoveTowards(double current, double target, double maxDelta)
+        static public float MoveTowards(float current, float target, float maxDelta)
         {
             if (Math.Abs(target - current) <= maxDelta)
                 return target;
@@ -233,9 +233,9 @@ namespace Electron2D
         }
 
         // Same as ::ref::MoveTowards but makes sure the values interpolate correctly when they wrap around 360 degrees.
-        static public double MoveTowardsAngle(double current, double target, double maxDelta)
+        static public float MoveTowardsAngle(float current, float target, float maxDelta)
         {
-            double deltaAngle = DeltaAngle(current, target);
+            float deltaAngle = DeltaAngle(current, target);
             if (-maxDelta < deltaAngle && deltaAngle < maxDelta)
                 return target;
             target = current + deltaAngle;
@@ -243,7 +243,7 @@ namespace Electron2D
         }
 
         // Interpolates between /min/ and /max/ with smoothing at the limits.
-        public static double SmoothStep(double from, double to, double t)
+        public static float SmoothStep(float from, float to, float t)
         {
             t = Math.Clamp01(t);
             t = (-2 * t * t * t) + (3 * t * t);
@@ -251,14 +251,14 @@ namespace Electron2D
         }
 
         //*undocumented
-        public static double Gamma(double value, double absmax, double gamma)
+        public static float Gamma(float value, float absmax, float gamma)
         {
             bool negative = value < 0F;
-            double absval = Abs(value);
+            float absval = Abs(value);
             if (absval > absmax)
                 return negative ? -absval : absval;
 
-            double result = Pow(absval / absmax, gamma) * absmax;
+            float result = Pow(absval / absmax, gamma) * absmax;
             return negative ? -result : result;
         }
 
@@ -275,42 +275,42 @@ namespace Electron2D
             return Abs(b - a) < Max(0.000001f * Max(Abs(a), Abs(b)), Epsilon * 8);
         }
 
-        public static double SmoothDamp(double current, double target, ref double currentVelocity, double smoothTime, double maxSpeed)
+        public static float SmoothDamp(float current, float target, ref float currentVelocity, float smoothTime, float maxSpeed)
         {
-            double deltaTime = Time.DeltaTime;
+            float deltaTime = Time.DeltaTime;
             return SmoothDamp(current, target, ref currentVelocity, smoothTime, maxSpeed, deltaTime);
         }
 
-        public static double SmoothDamp(double current, double target, ref double currentVelocity, double smoothTime)
+        public static float SmoothDamp(float current, float target, ref float currentVelocity, float smoothTime)
         {
-            double deltaTime = Time.DeltaTime;
-            const double maxSpeed = Math.Infinity;
+            float deltaTime = Time.DeltaTime;
+            const float maxSpeed = Math.Infinity;
             return SmoothDamp(current, target, ref currentVelocity, smoothTime, maxSpeed, deltaTime);
         }
 
         // Gradually changes a value towards a desired goal over time.
-        public static double SmoothDamp(double current, double target, ref double currentVelocity, double smoothTime,  double maxSpeed, double deltaTime)
+        public static float SmoothDamp(float current, float target, ref float currentVelocity, float smoothTime,  float maxSpeed, float deltaTime)
         {
             // Based on Game Programming Gems 4 Chapter 1.10
-            smoothTime = Math.Max(0.0001, smoothTime);
-            double omega = 2 / smoothTime;
+            smoothTime = Math.Max(0.0001f, smoothTime);
+            float omega = 2f / smoothTime;
 
-            double x = omega * deltaTime;
-            double exp = 1 / (1 + x + (0.48 * x * x) + (0.235 * x * x * x));
-            double change = current - target;
-            double originalTo = target;
+            float x = omega * deltaTime;
+            float exp = 1f / (1f + x + (0.48f * x * x) + (0.235f * x * x * x));
+            float change = current - target;
+            float originalTo = target;
 
             // Clamp maximum speed
-            double maxChange = maxSpeed * smoothTime;
+            float maxChange = maxSpeed * smoothTime;
             change = Math.Clamp(change, -maxChange, maxChange);
             target = current - change;
 
-            double temp = (currentVelocity + (omega * change)) * deltaTime;
+            float temp = (currentVelocity + (omega * change)) * deltaTime;
             currentVelocity = (currentVelocity - (omega * temp)) * exp;
-            double output = target + ((change + temp) * exp);
+            float output = target + ((change + temp) * exp);
 
             // Prevent overshooting
-            if (originalTo - current > 0 == output > originalTo)
+            if (originalTo - current > 0F == output > originalTo)
             {
                 output = originalTo;
                 currentVelocity = (output - originalTo) / deltaTime;
@@ -319,41 +319,41 @@ namespace Electron2D
             return output;
         }
 
-        public static double SmoothDampAngle(double current, double target, ref double currentVelocity, double smoothTime, double maxSpeed)
+        public static float SmoothDampAngle(float current, float target, ref float currentVelocity, float smoothTime, float maxSpeed)
         {
-            double deltaTime = Time.DeltaTime;
+            float deltaTime = Time.DeltaTime;
             return SmoothDampAngle(current, target, ref currentVelocity, smoothTime, maxSpeed, deltaTime);
         }
 
-        public static double SmoothDampAngle(double current, double target, ref double currentVelocity, double smoothTime)
+        public static float SmoothDampAngle(float current, float target, ref float currentVelocity, float smoothTime)
         {
-            double deltaTime = Time.DeltaTime;
-            const double maxSpeed = Math.Infinity;
+            float deltaTime = Time.DeltaTime;
+            const float maxSpeed = Math.Infinity;
             return SmoothDampAngle(current, target, ref currentVelocity, smoothTime, maxSpeed, deltaTime);
         }
 
         // Gradually changes an angle given in degrees towards a desired goal angle over time.
-        public static double SmoothDampAngle(double current, double target, ref double currentVelocity, double smoothTime, double maxSpeed, double deltaTime)
+        public static float SmoothDampAngle(float current, float target, ref float currentVelocity, float smoothTime, float maxSpeed, float deltaTime)
         {
             target = current + DeltaAngle(current, target);
             return SmoothDamp(current, target, ref currentVelocity, smoothTime, maxSpeed, deltaTime);
         }
 
         // Loops the value t, so that it is never larger than length and never smaller than 0.
-        public static double Repeat(double t, double length)
+        public static float Repeat(float t, float length)
         {
             return Clamp(t - (Math.Floor(t / length) * length), 0, length);
         }
 
         // PingPongs the value t, so that it is never larger than length and never smaller than 0.
-        public static double PingPong(double t, double length)
+        public static float PingPong(float t, float length)
         {
             t = Repeat(t, length * 2);
             return length - Math.Abs(t - length);
         }
 
         // Calculates the ::ref::Lerp parameter between of two values.
-        public static double InverseLerp(double a, double b, double value)
+        public static float InverseLerp(float a, float b, float value)
         {
             if (a != b)
                 return Clamp01((value - a) / (b - a));
@@ -362,9 +362,9 @@ namespace Electron2D
         }
 
         // Calculates the shortest difference between two given angles.
-        public static double DeltaAngle(double current, double target)
+        public static float DeltaAngle(float current, float target)
         {
-            double delta = Math.Repeat((target - current), 360);
+            float delta = Math.Repeat(target - current, 360);
             if (delta > 180)
                 delta -= 360;
             return delta;
@@ -373,18 +373,18 @@ namespace Electron2D
         // Infinite Line Intersection (line1 is p1-p2 and line2 is p3-p4)
         internal static bool LineIntersection(Vector p1, Vector p2, Vector p3, Vector p4, ref Vector result)
         {
-            double bx = p2.X - p1.X;
-            double by = p2.Y - p1.Y;
-            double dx = p4.X - p3.X;
-            double dy = p4.Y - p3.Y;
-            double bDotDPerp = (bx * dy) - (by * dx);
+            float bx = p2.X - p1.X;
+            float by = p2.Y - p1.Y;
+            float dx = p4.X - p3.X;
+            float dy = p4.Y - p3.Y;
+            float bDotDPerp = (bx * dy) - (by * dx);
             if (bDotDPerp == 0)
             {
                 return false;
             }
-            double cx = p3.X - p1.X;
-            double cy = p3.Y - p1.Y;
-            double t = ((cx * dy) - (cy * dx)) / bDotDPerp;
+            float cx = p3.X - p1.X;
+            float cy = p3.Y - p1.Y;
+            float t = ((cx * dy) - (cy * dx)) / bDotDPerp;
 
             result.X = p1.X + (t * bx);
             result.Y = p1.Y + (t * by);
@@ -394,23 +394,23 @@ namespace Electron2D
         // Line Segment Intersection (line1 is p1-p2 and line2 is p3-p4)
         internal static bool LineSegmentIntersection(Vector p1, Vector p2, Vector p3, Vector p4, ref Vector result)
         {
-            double bx = p2.X - p1.X;
-            double by = p2.Y - p1.Y;
-            double dx = p4.X - p3.X;
-            double dy = p4.Y - p3.Y;
-            double bDotDPerp = bx * dy - by * dx;
+            float bx = p2.X - p1.X;
+            float by = p2.Y - p1.Y;
+            float dx = p4.X - p3.X;
+            float dy = p4.Y - p3.Y;
+            float bDotDPerp = bx * dy - by * dx;
             if (bDotDPerp == 0)
             {
                 return false;
             }
-            double cx = p3.X - p1.X;
-            double cy = p3.Y - p1.Y;
-            double t = ((cx * dy) - (cy * dx)) / bDotDPerp;
+            float cx = p3.X - p1.X;
+            float cy = p3.Y - p1.Y;
+            float t = ((cx * dy) - (cy * dx)) / bDotDPerp;
             if (t < 0 || t > 1)
             {
                 return false;
             }
-            double u = ((cx * by) - (cy * bx)) / bDotDPerp;
+            float u = ((cx * by) - (cy * bx)) / bDotDPerp;
             if (u < 0 || u > 1)
             {
                 return false;
